refactor(gcp): restructure scaffolder backend module
- Move files from root directory to `src` folder for better organization
- Remove example action and tests
- Update module ID from `example-action` to `scaffolder-backend-module-aws`
- Update copyright notice to Datolabs, MB
- Add GCP Secrets Manager action implementation
