--- workspaces/scaffolder-backend-module-gcp/plugins/scaffolder-backend-module-gcp/module.ts
@@ -1,38 +1,0 @@
-/*
- * Copyright 2024 Datolabs, MB
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { createBackendModule } from '@backstage/backend-plugin-api';
-import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
-import * as backendModuleGCP from './actions';
-
-/**
- * @public
- */
-export const scaffolderBackendModuleGCP = createBackendModule({
-  pluginId: 'scaffolder',
-  moduleId: 'scaffolder-backend-module-aws',
-  register({ registerInit }) {
-    registerInit({
-      deps: {
-        scaffolder: scaffolderActionsExtensionPoint,
-      },
-      async init({ scaffolder }) {
-        scaffolder.addActions(
-          backendModuleGCP.createGcpSecretsManagerCreateAction(),
-        );
-      },
-    });
-  },
-});
