+++ workspaces/scaffolder-backend-module-gcp/plugins/scaffolder-backend-module-gcp/src/actions/secrets-manager/index.ts
