+++ workspaces/scaffolder-backend-module-gcp/plugins/scaffolder-backend-module-gcp/src/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 The Backstage Authors
+ * Copyright 2024 Datolabs, MB
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-/**
- * The gcp module for @backstage/plugin-scaffolder-backend.
- *
- * @packageDocumentation
- */
-
-export { scaffolderModule as default } from './module';
+export * from './module';
+export { scaffolderBackendModuleGCP as default } from './module';
